==> R/06_ancestral_biogeo_model.R <==
####### Writing began in August 2022 by Thales
####### Written during the course Introduction to Scientific Computation
####### Written as part of the final evaluation of this same course
####### Written based on the template script of the package BioGeoBEARS
####### Serrapilheira ICTP/SAIRF QBIO program

####### This script uses the biogeographic information generated by script 05
####### as an input in a ancestral range estimation biogeographic model.
####### The model used is an adapted likelihood version of
####### the DIVA (dispersal-vicariance) model (Ronquist (1997).

####### The whole implementation is done using the BioGeoBEARS package (Matzke, Nicholas J. (2018))
####### The base script of the package, used as template to this one, can be accessed at:
####### http://phylo.wikidot.com/biogeobears#script
####### This site provides also a comprehensive explanation of the package and additional
####### relevant information
####### Other relevant information can be found at the GitHub page of the package:
####### https://github.com/nmatzke/BioGeoBEARS



### Installing the BioGeoBEARS ------------------------------------------------------
### Based at in the tutorial available at the package's GitHub page

# Install the rexpokit and cladoRcpp packages, both available on CRAN
install.packages("rexpokit")
install.packages("cladoRcpp")

# Install the new version of BioGeoBEARS from GitHub, using devtools:
library(devtools)
devtools::install_github(repo="nmatzke/BioGeoBEARS")
# When it asks which library do you want to update, select none (usaully option #3)

# If step #2 keeps trying to reinstall rexpokit and cladoRcpp from scratch (including
# demanding a compiler like gcc or gfortran) even though they are already
# installed, try adding "dependencies=FALSE":
 devtools::install_github(repo="nmatzke/BioGeoBEARS", dependencies=FALSE)

# If you hit other needed dependencies, install them individually from CRAN,
# then re-run the command above.




### Library -----------------------------------------------------------------
library(optimx)
library(GenSA)
library(FD)
library(snow)
library(parallel)
library(ape)

library(rexpokit)
library(cladoRcpp)
library(BioGeoBEARS)

source("fun/01_DIVALIKE.R") # see the function files for details



### Importing the phylogenetic tree in Newick format ----------------------------
trfn <-  "./output/ultra_rerooted_coregroup_tree.newick"

# Look at the raw Newick file
moref(trfn)

# Read the tree
tr <- read.tree(trfn)



### Importing the file with the biogeographic data -------------------------------
# This file is an manually edited version of the output/spp_bio_reg_BEARS.txt
# Generated by the script 05_spp_biogeo_obtention.R
# The edition was made to set the file to the pattern need by BioGeoBEARS package
geogfn <- "./output/spp_bio_reg_BEARS.data"

# Look at the raw geography text file:
moref(geogfn)

# Better visualization of the biogeographic file
tipranges <- getranges_from_LagrangePHYLIP(lgdata_fn=geogfn)
tipranges

# Which is the maximal geographic span (in number of biogeographic regions) in the data set
max_observed <- max(rowSums(dfnums_to_numeric(tipranges@df)))

# Setting the maximum range of occupied areas in the reconstruction as the maximum
# number of occupied areas by extant taxa

resDIVALIKE <- DIVALIKE(trfn = trfn, geogfn = geogfn, max_range_size = max_observed)


resDIVALIKEj <- DIVALIKE_J(trfn = trfn, geogfn = geogfn, max_range_size = max_observed,
                           resDIVALIKE = resDIVALIKE)




#### Comparing models through AIC ---------------------------------------------------------

n_par <- c(3, 2) # the DIVALIKE has 2 parameters and the DIVALIKE+J has 3

# Models negative log likelihood
DIVALIKE <- resDIVALIKE[["optim_result"]][["value"]]
DIVALIKEj<- resDIVALIKEj[["optim_result"]][["value"]]

LnL <- rbind(DIVALIKEj, DIVALIKE)

# AIC calculation
AIC <- (2*n_par) - (2*LnL)
delta_AIC <- abs(AIC - min(AIC))
AIC_table <- cbind(n_par, LnL, AIC, delta_AIC)

colnames(AIC_table)[4] <- "delta AIC"
colnames(AIC_table)[3] <- "AIC"
colnames(AIC_table)[2] <- "LnL"
colnames(AIC_table)[1] <- "K"
head(AIC_table)

# Saving AIC table
write.table(AIC_table, "./output/model_AIC.txt")




#### Plotting the best model ------------------------------------------------------------

# Color of the areas
areas_col <- c("black", "blue", "lawngreen", "red", "yellow",
               "cyan2", "purple", "green", "orange", "grey3", "grey")


# Plot ancestral states - DIVALIKE+J in pdf
pdffn = "./figs/Calamoideae_DIVALIKE+J_M0_unconstrained_v1.pdf"
pdf(pdffn, width=12, height=8)
analysis_titletxt ="BioGeoBEARS DIVALIKE+J on Calamoideae M0_unconstrained"

# Setup
results_object = resDIVALIKEj
scriptdir = np(system.file("extdata/a_scripts", package="BioGeoBEARS"))

# Pie chart
plot_BioGeoBEARS_results(results_object, analysis_titletxt,
                         addl_params= F, plotwhat="pie",
                         label.offset=0.02, tipcex=0.9, statecex=0.4,
                         titlecex=1, plotsplits = F,
                         cornercoords_loc=scriptdir,
                         include_null_range=TRUE, tr=tr,
                         tipranges=tipranges,
                         pie_tip_statecex = 0.1, tiplabel_adj  = -2,
                         plotlegend = F,
                         colors_list_for_states = areas_col, xlims=NULL,
                         xlab = "")

dev.off()
cmdstr = paste("open ", pdffn, sep="")
system(cmdstr)

graphics.off()
